requirements.txt
transformers==4.35.0
sentence-transformers==2.2.2
torch==2.2.0
chromadb==0.4.22
PyMuPDF==1.23.8
python-docx==1.1.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.11.7
loguru==0.7.2
tqdm==4.66.1
nltk==3.8.1
spacy==3.7.2
python-dotenv==1.0.1
pdfplumber==0.10.3
httpx==0.27.0
pytest==8.1.1 